+++ TypedSyntax/Project.toml
@@ -7,12 +7,6 @@ version = "1.2.1"
 CodeTracking = "da1fd8a2-8d9e-5ec2-8556-3022fb5608a2"
 JuliaSyntax = "70703baa-626e-46a2-a12c-08ffd08c73b4"
 
-[weakdeps]
-VSCodeServer = "9f5989ce-84fe-42d4-91ec-6a7a8d53ed0f"
-
-[extensions]
-VSCodeExt = "VSCodeServer"
-
 [compat]
 CodeTracking = "1.3"
 JuliaSyntax = "0.4"
@@ -25,7 +19,6 @@ Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 MethodAnalysis = "85b6ec6f-f7df-4429-9514-a64bcd9ee824"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-VSCodeServer = "9f5989ce-84fe-42d4-91ec-6a7a8d53ed0f"
 
 [targets]
 test = ["Dates", "InteractiveUtils", "Logging", "MethodAnalysis", "ProgressMeter", "Test"]
