+++ TypedSyntax/src/TypedSyntax.jl
@@ -12,5 +12,6 @@ export TypedSyntaxNode
 
 include("node.jl")
 include("show.jl")
+include("vscode.jl")
 
 end
